requirements.txt
pandas
pyarrow
scikit-learn
pyyaml
joblib